pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-annotator"
version = "0.1.0"
description = "Media Annotator & Smart Renamer"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Media Annotator"}]
dependencies = [
  "pydantic>=2.5",
  "typer>=0.12",
  "rich>=13.7",
  "sqlalchemy>=2.0",
  "opencv-python>=4.8",
  "pillow>=10.0",
  "numpy>=1.26",
  "python-dateutil>=2.9",
  "loguru>=0.7",
  "xxhash>=3.4",
  "httpx>=0.27",
]

[project.optional-dependencies]
faces = [
  "insightface>=0.7.3",
  "onnxruntime>=1.16",
]
llm_ollama = [
  "httpx>=0.27",
]
llm_lmstudio = [
  "openai>=1.30",
]
llm_local = [
  "torch>=2.1",
  "transformers>=4.40",
  "accelerate>=0.27",
  "safetensors>=0.4",
]
gui = [
  "PySide6>=6.6",
  "qdarktheme>=2.1",
]
faiss = [
  "faiss-cpu>=1.8",
]

[project.scripts]
media-annotator = "media_annotator.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["media_annotator*"]

[tool.setuptools]
include-package-data = true
